requirements.txt
plotly==5.14.1
haversine==2.8.0
folium==0.14.0
streamlit==1.21.0
streamlit-folium==0.11.1
pandas==1.5.3
numpy==1.24.2
Pillow==9.4.0
